requirements.txt
fastapi
uvicorn
spacy
python-dotenv
